Add promised `oneEvent` listener
