+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
-index.js
-index.d.ts
+*.js
+!*.test.js
+*.d.ts
 *.map
